pyproject.toml
[tool.poetry]
name = "bluenaas"
version = "0.1.0"
description = ""
authors = ["Bilal Meddah <bilal.meddah@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
bluecellulab = "^2.6.35"
ruff = "^0.5.2"
mypy = "^1.10.1"
loguru = "^0.7.2"
pydantic = "^2.8.2"
requests = "^2.32.3"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
python-keycloak = "4.3.0"
sympy = "^1.13.1"
pandas = "^2.2.2"
annotated-types = "^0.7.0"
sentry-sdk = {version = "^2.13.0", extras = ["fastapi"]}
filelock = "^3.16.0"
celery = {extras = ["redis", "sqlalchemy"], version = "^5.4.0"}
boto3 = "^1.35.28"
psycopg2-binary = "^2.9.9"
sqlalchemy = "2.0.35"
redis = "5.1.0"
psutil = "^6.0.0"
watchdog = {extras = ["watchmedo"], version = "^5.0.3"}
billiard = "^4.2.1"
mypy-boto3 = "^1.35.34"
boto3-stubs = {extras = ["cloudwatch", "ec2", "ecs"], version = "^1.35.34"}
scalar-fastapi = "^1.0.3"

[tool.poetry.group.types.dependencies]
types-requests = "^2.32.0.20240712"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "/Users/meddah/Library/Caches/pypoetry/virtualenvs"
venv = "bluenaas-DrmS_1CG-py3.12"

[tool.poetry.scripts]
monitor = "bluenaas.infrastructure.celery.events_monitoring:main"